pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "window-aichat"
version = "2.0.0"
description = "A desktop AI chat application with GitHub integration"
authors = [
    {name = "Developer", email = "dev@example.com"},
]
dependencies = [
    "google-generativeai>=0.3.0",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "cryptography",
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-multipart",
    "aiofiles",
    "psutil",
    "pygments",
    "httpx",
    "sqlalchemy",
    "alembic",
    "python-jose[cryptography]",
    "python-dotenv",
    "tiktoken"
]
requires-python = ">=3.8"

[project.scripts]
window-aichat = "window_aichat.__main__:main"
